requirements.txt
langchain
langchain-community
langchain-groq
chromadb
streamlit
pandas
python-dotenv
langchain_huggingface  
sentence-transformers 

